Move investment settings to own tab
