+++ src/components/Settings/Investment/Investment.css
@@ -1,0 +1,7 @@
+.Investment{
+  height: 60vh;
+}
+
+.btn-reinvest{
+  margin: .2rem auto;
+}
